feat(vs2022_third): C4477 fix final remaining warnings — 7 files
Fix remaining format specifier mismatches:
- Recipe*ListView files (5 files): Add (LPCTSTR) casts for CString→char*
- ChiefView.cpp: Cast double to int for %d format
- STAGE.H: Use local variable for template parameter to satisfy C4477

Total: 16 remaining C4477 warnings fixed.
No behavior changes, only explicit casts and local variable to match format specifiers.

+++ SRC/NanoSpec/ChiefView.cpp
@@ -2583,7 +2583,7 @@ BOOL CChiefView::SelectSubstrateThick(double dVal)
 	ActuateFlagsSet(ACTUATE_ZAXIS, FALSE);
 
 	TCHAR l_tszLogText[256];
-	_stprintf(l_tszLogText, _T("Selected Substrate Thickness(=%d), Result(=%d)"), dVal, l_bRc);
+	_stprintf(l_tszLogText, _T("Selected Substrate Thickness(=%d), Result(=%d)"), (int)dVal, l_bRc);
 	LogChief(l_tszLogText);
 
 	if (0 == l_bRc) {
